Last commit as  of today

+++ src/Controller/UserController.php
@@ -10,7 +10,6 @@ use Symfony\Bundle\FrameworkBundle\Controller\AbstractController;
 use Symfony\Component\HttpFoundation\Request;
 use Symfony\Component\HttpFoundation\Response;
 use Symfony\Component\Routing\Attribute\Route;
-use Symfony\Component\Security\Core\User\UserInterface;
 
 #[Route('/user')]
 final class UserController extends AbstractController
@@ -18,17 +17,25 @@ final class UserController extends AbstractController
     #[Route(name: 'app_user_index', methods: ['GET'])]
     public function index(UserRepository $userRepository): Response
     {
+        if(!$this->isGranted('ROLE_USER')){
+            throw $this->createAccessDeniedException('Merci de vous connecter');}
+            else{
         return $this->render('user/index.html.twig', [
             'users' => $userRepository->findAll(),
         ]);
+            }
     }
 
     #[Route('/new', name: 'app_user_new', methods: ['GET', 'POST'])]
     public function new(Request $request, EntityManagerInterface $entityManager): Response
     {
+        if(!$this->isGranted('ROLE_ADMIN')){
+            throw $this->createAccessDeniedException('Accès réservé à un administrateur');}
+            else{
         $user = new User();
         $form = $this->createForm(UserType::class, $user);
         $form->handleRequest($request);
+        }
 
         if ($form->isSubmitted() && $form->isValid()) {
             $entityManager->persist($user);
@@ -41,19 +48,26 @@ final class UserController extends AbstractController
             'user' => $user,
             'form' => $form,
         ]);
-    }
+    } 
 
     #[Route('/{id}', name: 'app_user_show', methods: ['GET'])]
     public function show(User $user): Response
     {
+        if(!$this->isGranted('ROLE_USER')){
+            throw $this->createAccessDeniedException('Merci de vous connecter');}
+            else{
         return $this->render('user/show.html.twig', [
             'user' => $user,
         ]);
+            }
     }
 
     #[Route('/{id}/edit', name: 'app_user_edit', methods: ['GET', 'POST'])]
     public function edit(Request $request, User $user, EntityManagerInterface $entityManager): Response
     {
+        if(!$this->isGranted('ROLE_USER')){
+            throw $this->createAccessDeniedException('Merci de vous connecter');}
+            else{
         $form = $this->createForm(UserType::class, $user);
         $form->handleRequest($request);
 
@@ -67,16 +81,21 @@ final class UserController extends AbstractController
             'user' => $user,
             'form' => $form,
         ]);
+            }
     }
 
     #[Route('/{id}', name: 'app_user_delete', methods: ['POST'])]
     public function delete(Request $request, User $user, EntityManagerInterface $entityManager): Response
     {
+        if(!$this->isGranted('ROLE_ADMIN')){
+            throw $this->createAccessDeniedException('Accès réservé à un administrateur');}
+            else{
         if ($this->isCsrfTokenValid('delete'.$user->getId(), $request->getPayload()->getString('_token'))) {
             $entityManager->remove($user);
             $entityManager->flush();
         }
 
         return $this->redirectToRoute('app_user_index', [], Response::HTTP_SEE_OTHER);
+    }
     }
 }
